daemon/logger: use uniform location for registering drivers
Logging-drivers are registered through an init function, but they were
burried deep in the implementation files; move the driver-registration
call to a separate file, consistent between all drivers.
